+++ components/index.ts
@@ -1,7 +1,7 @@
 export { default as Button } from './Button'
 export { default as Carousel } from './Carousel'
 export { default as CheckConnectedWallet } from './CheckConnectedWallet'
-export { default as Modal } from './Modal'
+export { default as Popup } from './Popup'
 export { default as Navbar } from './Navbar'
 export { default as Spinner } from './Spinner'
 export { default as Title } from './Title'
